created singleLocationReducer module and corrispetive test file
