+++ src/redux/reducers/__tests__/locationsReducer.test.js
