+++ src/redux/reducers/reducers.test.js
@@ -1,79 +1,6 @@
 import { singleLocaReducer, mapsReducer } from './reducers';
 import * as types from '../actionTypes';
 
-describe('single location reducer', () => {
-  const location = {
-    id: 1,
-    description: '123 Arizona road',
-    name: 'Ariz Coffee Shop',
-    longitude: '0.999923',
-    latitude: '0.273444',
-  };
-
-  it('should return the initial state', () => {
-    expect(singleLocaReducer(undefined, {})).toEqual({
-      loadingSingleLoc: false,
-      location: null,
-      error: null,
-    });
-  });
-  it('should load locations', () => {
-    expect(
-      singleLocaReducer(
-        { loadingSingleLoc: false },
-        {
-          type: types.LOADING_SINGLE_LOCATION,
-        },
-      ),
-    ).toEqual({
-      loadingSingleLoc: true,
-    });
-    // test the spread state functionality
-    expect(
-      singleLocaReducer(
-        { loadingSingleLoc: false, location: null, error: null },
-        {
-          type: types.LOADING_SINGLE_LOCATION,
-        },
-      ),
-    ).toEqual({
-      loadingSingleLoc: true,
-      location: null,
-      error: null,
-    });
-  });
-  it('should get a single locations on success', () => {
-    expect(
-      singleLocaReducer(
-        { loadingSingleLoc: false, location: null, error: null },
-        {
-          type: types.FETCH_SINGLE_LOCATIONS_SUCCESS,
-          payload: location,
-        },
-      ),
-    ).toEqual({
-      loadingSingleLoc: false,
-      location,
-      error: null,
-    });
-  });
-  it('should show error on failure', () => {
-    expect(
-      singleLocaReducer(
-        { loadingSingleLoc: false, location: null, error: null },
-        {
-          type: types.FETCH_SINGLE_LOCATIONS_FAILURE,
-          payload: 'There was an error',
-        },
-      ),
-    ).toEqual({
-      loadingSingleLoc: false,
-      location: null,
-      error: 'There was an error',
-    });
-  });
-});
-
 describe('geolocation reducer', () => {
 
   it('should return the initial state', () => {
