+++ src/redux/reducers/rootReducer.js
@@ -1,12 +1,13 @@
 import { combineReducers } from 'redux';
 import * as reducers from './reducers';
 import { locationsReducer } from './locationsReducer'
+import { singleLocationReducer } from './singleLocationReducer'
 
 const rootReducer = combineReducers({
   locations: locationsReducer,
   newLocation: reducers.addLocationReducer,
   maps: reducers.mapsReducer,
-  location: reducers.singleLocaReducer,
+  location: singleLocationReducer,
   activeLocation: reducers.activeLocation,
   auth: reducers.authReducer,
   verifyEmail: reducers.verifyEmailReducer,
